Relocate tir stuff to `baml_type`
This is in preparation for larger changes. Also adds `subenum` for
`baml_type::Ty` (will be extended by further changes)
