+++ baml_language/Cargo.toml
@@ -200,6 +200,7 @@ similar = { version = "2.4.0", features = [ "inline" ] }
 smallvec = { version = "1.15", features = [ "const_generics" ] }
 smol_str = { version = "0.3", features = [ "serde", "borsh" ] }
 strsim = { version = "0.11.1" }
+subenum = { version = "1.2.0" }
 syn = { version = "2", features = [ "full" ] }
 tar = "0.4"
 zip = { version = "2", default-features = false, features = [ "deflate" ] }
