+++ baml_language/crates/baml_compiler2_tir/src/ty.rs
@@ -4,240 +4,14 @@ use std::fmt;
 
 use baml_base::Name;
 pub use baml_base::attr::TyAttr;
-
-/// A qualified type name with separate package and local name.
-///
-/// Used in `Ty::Class`, `Ty::Enum`, and `Ty::TypeAlias` to unambiguously
-/// identify a type by its definition's package (e.g. `"user"`, `"baml"`)
-/// and its short name (e.g. `"Foo"`, `"PrimitiveClient"`).
-/// Which package a type is defined in. `Local` is the user's own (implicit
-/// root) package — the "current" package for everything a user writes;
-/// `Dep(name)` is a named dependency (e.g. `baml`). Encoding this as a type
-/// rather than a magic `"user"` string means the local-vs-dependency
-/// distinction is checked by the compiler, not by string comparison: the only
-/// place the `"user"` string appears is [`Package::from_name`] (the boundary
-/// where upstream `Name`-based package info is classified).
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub enum Package {
-    /// The user's own implicit root package (`RESERVED_USER_PACKAGE`).
-    Local,
-    /// A named dependency package.
-    Dep(Name),
-}
-
-/// The interned `Name` of the reserved implicit `user` package, materialized
-/// once so [`QualifiedTypeName::package`] can hand out a `&Name` for `Local`.
-static USER_PACKAGE_NAME: Name = Name::new_inline(RESERVED_USER_PACKAGE);
-
-impl Package {
-    /// Classify an upstream package `Name`: the reserved `user` package becomes
-    /// [`Package::Local`], everything else a [`Package::Dep`]. This is the one
-    /// spot the `"user"` magic string is read.
-    pub fn from_name(name: Name) -> Self {
-        if name.as_str() == RESERVED_USER_PACKAGE {
-            Package::Local
-        } else {
-            Package::Dep(name)
-        }
-    }
-
-    /// The package's `Name` (`Local` resolves to the reserved `user` name).
-    pub fn as_name(&self) -> &Name {
-        match self {
-            Package::Local => &USER_PACKAGE_NAME,
-            Package::Dep(name) => name,
-        }
-    }
-}
-
-// Order/sort by the package *name* string, preserving the pre-enum `Ord`
-// (where `pkg` was a `Name`) so `QualifiedTypeName`'s derived ordering — and
-// any sorted output keyed on it — is unchanged.
-impl Ord for Package {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.as_name().as_str().cmp(other.as_name().as_str())
-    }
-}
-
-impl PartialOrd for Package {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub struct QualifiedTypeName {
-    /// The package this type is defined in (`Local` for user code, `Dep` for a
-    /// dependency like `baml`).
-    pkg: Package,
-    /// The namespace this type is defined in (e.g. `["llm"]`).
-    namespace: Vec<Name>,
-    /// The short/local name of the type (e.g. `"Foo"`).
-    name: Name,
-    /// Unresolved generic type parameter names (e.g. `["T"]` for `Array<T>`).
-    /// Empty for non-generic types or when concrete type args are substituted.
-    pub generic_params: Vec<Name>,
-}
-
-impl QualifiedTypeName {
-    pub fn new(pkg: Name, namespace: Vec<Name>, name: Name) -> Self {
-        Self::new_with_generic_params(pkg, namespace, name, Vec::new())
-    }
-
-    pub fn new_with_generic_params(
-        pkg: Name,
-        namespace: Vec<Name>,
-        name: Name,
-        generic_params: Vec<Name>,
-    ) -> Self {
-        Self {
-            pkg: Package::from_name(pkg),
-            namespace,
-            name,
-            generic_params,
-        }
-    }
-
-    pub fn package(&self) -> &Name {
-        self.pkg.as_name()
-    }
-
-    /// Whether this type lives in the user's own (implicit root) package — the
-    /// "current" package for everything a user writes. User-facing rendering
-    /// omits the package for these; only dependency types carry a package
-    /// qualifier. Use this instead of comparing `package()` to `"user"`.
-    pub fn is_local(&self) -> bool {
-        matches!(self.pkg, Package::Local)
-    }
-
-    pub fn namespace(&self) -> &Vec<Name> {
-        &self.namespace
-    }
-
-    pub fn name(&self) -> &Name {
-        &self.name
-    }
-
-    pub fn is_builtin_root_type(&self, name: &str) -> bool {
-        self.package().as_str() == "baml" && self.namespace.is_empty() && self.name.as_str() == name
-    }
-
-    /// Returns `true` if this type lives in the `baml.panics` namespace
-    /// (i.e. it is a panic class or the `Panic` type alias).
-    pub fn is_panic_type(&self) -> bool {
-        baml_base::is_panic_namespace(self.package().as_str(), &self.namespace)
-    }
-
-    pub fn to_path_in_package(&self) -> Vec<Name> {
-        self.namespace
-            .iter()
-            .chain(std::iter::once(&self.name))
-            .cloned()
-            .collect::<Vec<_>>()
-    }
-
-    /// The dotted path `package.namespace.name` (no `<generic_params>` suffix).
-    /// When `user_facing`, the reserved implicit `user` package is elided
-    /// ([`RESERVED_USER_PACKAGE`]) — the single structural source of the
-    /// "no `user.` in names" rule. The canonical form (`user_facing = false`)
-    /// keeps the package for dumps/identity.
-    pub fn render_dotted(&self, user_facing: bool) -> String {
-        let namespace = self
-            .namespace
-            .iter()
-            .map(std::string::ToString::to_string)
-            .collect::<Vec<_>>()
-            .join(".");
-        let elide = user_facing && self.is_local();
-        let pkg = self.package();
-        match (elide, namespace.is_empty()) {
-            (true, true) => self.name.to_string(),
-            (true, false) => format!("{namespace}.{}", self.name),
-            (false, true) => format!("{}.{}", pkg, self.name),
-            (false, false) => format!("{}.{namespace}.{}", pkg, self.name),
-        }
-    }
-
-    /// Like [`render_dotted`](Self::render_dotted) plus the declared
-    /// `<generic_params>` suffix (e.g. `Array<T>`).
-    pub fn render_qualified(&self, user_facing: bool) -> String {
-        let mut out = self.render_dotted(user_facing);
-        if !self.generic_params.is_empty() {
-            let params: Vec<_> = self
-                .generic_params
-                .iter()
-                .map(std::string::ToString::to_string)
-                .collect();
-            out.push('<');
-            out.push_str(&params.join(", "));
-            out.push('>');
-        }
-        out
-    }
-
-    /// User-facing rendering of the qualified name: identical to the canonical
-    /// [`fmt::Display`] except the reserved implicit `user` package is elided.
-    /// Call this instead of post-processing the canonical string.
-    pub fn render_user_facing(&self) -> String {
-        self.render_qualified(true)
-    }
-
-    /// If this names a builtin `baml` companion class that has a lowercase
-    /// primitive/keyword alias, return that alias: `baml.String` → `string`,
-    /// `baml.media.Image` → `image`, `baml.json.json` → `json`. Returns `None`
-    /// for any other type (including user types and non-aliased `baml` types
-    /// such as `baml.json.JsonObject`).
-    ///
-    /// This is the single collapse rule used by the describe/hover canonical
-    /// type printer; it must stay in sync with
-    /// [`PrimitiveType::builtin_class_path`].
-    pub fn builtin_alias(&self) -> Option<&'static str> {
-        if self.package().as_str() != "baml" {
-            return None;
-        }
-        // `json` is the `baml.json.json` type alias, not a `PrimitiveType`.
-        if self.namespace.len() == 1
-            && self.namespace[0].as_str() == "json"
-            && self.name.as_str() == "json"
-        {
-            return Some("json");
-        }
-        let path: Vec<&str> = self
-            .namespace
-            .iter()
-            .map(Name::as_str)
-            .chain(std::iter::once(self.name.as_str()))
-            .collect();
-        PrimitiveType::from_builtin_class_path(&path).map(|p| p.alias())
-    }
-}
-
-/// The reserved implicit root package for user-authored code. It is the
-/// *current* package for everything a user writes, so it must never be shown in
-/// user-facing output (`user.Dog` → `Dog`). The canonical `Display` keeps it
-/// (for dumps/identity); only the user-facing path elides it.
-pub const RESERVED_USER_PACKAGE: &str = "user";
-
-/// Prefix of synthetic effect-polymorphism type parameters. These are an
-/// internal encoding (`__effect_param_0`, …); user-facing rendering shows them
-/// as `callback`. Single source of truth — use [`is_synthetic_effect_param`]
-/// rather than re-deriving this prefix check.
-pub const SYNTHETIC_EFFECT_PARAM_PREFIX: &str = "__effect_param_";
-
-/// Whether `name` is a synthesized effect-polymorphism type parameter
-/// (`__effect_param_N`). The single source of truth for this check — TIR, MIR,
-/// and the LSP all call here instead of re-implementing the prefix match.
-pub fn is_synthetic_effect_param(name: &Name) -> bool {
-    name.as_str()
-        .strip_prefix(SYNTHETIC_EFFECT_PARAM_PREFIX)
-        .is_some_and(|rest| !rest.is_empty() && rest.bytes().all(|b| b.is_ascii_digit()))
-}
-
-impl fmt::Display for QualifiedTypeName {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.render_qualified(false))
-    }
-}
+// `Package`, `QualifiedTypeName`, the reserved-package / synthetic-effect-param
+// constants, and `is_synthetic_effect_param` now live in `baml_type` (the
+// single home for the shared type vocabulary). Re-exported here so existing
+// `crate::ty::…` / `baml_compiler2_tir::ty::…` paths keep working.
+pub use baml_type::{
+    Freshness, Package, PrimitiveType, QualifiedTypeName, RESERVED_USER_PACKAGE,
+    SYNTHETIC_EFFECT_PARAM_PREFIX, is_synthetic_effect_param,
+};
 
 /// Resolved type — the output of type resolution (Pass 2).
 #[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
@@ -428,109 +202,6 @@ impl FunctionParamTy {
 pub enum FunctionParamMode {
     Required,
     Optional,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub enum PrimitiveType {
-    Int,
-    Bigint,
-    Float,
-    String,
-    Bool,
-    Null,
-    Uint8Array,
-    Image,
-    Audio,
-    Video,
-    Pdf,
-}
-
-impl PrimitiveType {
-    /// Map primitives with builtin companion classes to their class path in the `baml` package.
-    ///
-    /// Media primitives (`image`, `audio`, `video`, `pdf`) have corresponding
-    /// classes in `baml_builtins2/baml_std/baml/ns_media/media.baml`, and
-    /// `uint8array` has its class in `baml_builtins2/baml_std/baml/uint8array.baml`.
-    pub fn builtin_class_path(&self) -> &'static [&'static str] {
-        match self {
-            Self::Int => &["Int"],
-            Self::Bigint => &["Bigint"],
-            Self::Float => &["Float"],
-            Self::Bool => &["Bool"],
-            Self::Null => &["Null"],
-            Self::String => &["String"],
-            Self::Uint8Array => &["Uint8Array"],
-            Self::Image => &["media", "Image"],
-            Self::Audio => &["media", "Audio"],
-            Self::Video => &["media", "Video"],
-            Self::Pdf => &["media", "Pdf"],
-        }
-    }
-
-    pub fn from_literal(lit: &baml_base::Literal) -> Self {
-        match lit {
-            baml_base::Literal::Int(_) => Self::Int,
-            baml_base::Literal::Bigint(_) => Self::Bigint,
-            baml_base::Literal::Float(_) => Self::Float,
-            baml_base::Literal::String(_) => Self::String,
-            baml_base::Literal::Bool(_) => Self::Bool,
-        }
-    }
-
-    /// The lowercase primitive/keyword spelling for this type (`string`, `int`,
-    /// `image`, …). Single source of truth — the [`fmt::Display`] impl delegates
-    /// here.
-    pub fn alias(&self) -> &'static str {
-        match self {
-            Self::Int => "int",
-            Self::Bigint => "bigint",
-            Self::Float => "float",
-            Self::String => "string",
-            Self::Bool => "bool",
-            Self::Null => "null",
-            Self::Uint8Array => "uint8array",
-            Self::Image => "image",
-            Self::Audio => "audio",
-            Self::Video => "video",
-            Self::Pdf => "pdf",
-        }
-    }
-
-    /// Inverse of [`builtin_class_path`](Self::builtin_class_path): map a class
-    /// path (relative to the `baml` package, e.g. `["media", "Image"]`) back to
-    /// the primitive it is the companion class for.
-    pub fn from_builtin_class_path(path: &[&str]) -> Option<Self> {
-        const ALL: [PrimitiveType; 11] = [
-            PrimitiveType::Int,
-            PrimitiveType::Bigint,
-            PrimitiveType::Float,
-            PrimitiveType::Bool,
-            PrimitiveType::Null,
-            PrimitiveType::String,
-            PrimitiveType::Uint8Array,
-            PrimitiveType::Image,
-            PrimitiveType::Audio,
-            PrimitiveType::Video,
-            PrimitiveType::Pdf,
-        ];
-        ALL.into_iter().find(|p| p.builtin_class_path() == path)
-    }
-}
-
-/// Freshness flag for literal types.
-///
-/// Modeled after TypeScript's fresh/regular literal type distinction.
-/// - **Fresh**: produced by literal expressions (`1`, `"hello"`). Widens to
-///   the base primitive at mutable binding sites (`let x = 1` → `int`).
-/// - **Regular**: produced by type annotations (`let x: 1 = 1`) or contextual
-///   typing. Preserved through mutable bindings.
-///
-/// Freshness is **ignored** by the subtype checker — `Literal(1, Fresh)` and
-/// `Literal(1, Regular)` are structurally identical for assignability.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub enum Freshness {
-    Fresh,
-    Regular,
 }
 
 /// Re-export `baml_base::Literal` as `LiteralValue` for backward compatibility.
@@ -991,86 +662,5 @@ impl fmt::Display for Ty {
             "{}",
             self.render_with(&CanonicalTyRender { user_facing: false })
         )
-    }
-}
-
-impl fmt::Display for PrimitiveType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.alias())
-    }
-}
-
-#[cfg(test)]
-mod alias_tests {
-    use baml_base::Name;
-
-    use super::*;
-
-    #[test]
-    fn primitive_alias_class_path_roundtrips() {
-        for p in [
-            PrimitiveType::Int,
-            PrimitiveType::Bigint,
-            PrimitiveType::Float,
-            PrimitiveType::String,
-            PrimitiveType::Bool,
-            PrimitiveType::Null,
-            PrimitiveType::Uint8Array,
-            PrimitiveType::Image,
-            PrimitiveType::Audio,
-            PrimitiveType::Video,
-            PrimitiveType::Pdf,
-        ] {
-            let path = p.builtin_class_path();
-            assert_eq!(
-                PrimitiveType::from_builtin_class_path(path),
-                Some(p.clone()),
-                "round-trip failed for {p:?} via {path:?}"
-            );
-            // The alias matches the Display spelling.
-            assert_eq!(p.alias(), p.to_string());
-        }
-    }
-
-    #[test]
-    fn from_builtin_class_path_rejects_unknown() {
-        assert_eq!(PrimitiveType::from_builtin_class_path(&["Nope"]), None);
-        assert_eq!(
-            PrimitiveType::from_builtin_class_path(&["media", "Nope"]),
-            None
-        );
-    }
-
-    fn baml_qtn(namespace: &[&str], name: &str) -> QualifiedTypeName {
-        QualifiedTypeName::new(
-            Name::new("baml"),
-            namespace.iter().copied().map(Name::new).collect(),
-            Name::new(name),
-        )
-    }
-
-    #[test]
-    fn builtin_alias_collapses_primitive_classes() {
-        assert_eq!(baml_qtn(&[], "String").builtin_alias(), Some("string"));
-        assert_eq!(baml_qtn(&[], "Int").builtin_alias(), Some("int"));
-        assert_eq!(baml_qtn(&["media"], "Image").builtin_alias(), Some("image"));
-        assert_eq!(baml_qtn(&["media"], "Pdf").builtin_alias(), Some("pdf"));
-    }
-
-    #[test]
-    fn builtin_alias_handles_json_special_case() {
-        // `json` is the `baml.json.json` type alias, not a `PrimitiveType`.
-        assert_eq!(baml_qtn(&["json"], "json").builtin_alias(), Some("json"));
-        // A non-aliased `baml.json` type collapses to nothing.
-        assert_eq!(baml_qtn(&["json"], "JsonObject").builtin_alias(), None);
-    }
-
-    #[test]
-    fn builtin_alias_ignores_user_and_unaliased_types() {
-        // User-package `String` is never collapsed.
-        let user = QualifiedTypeName::new(Name::new("user"), vec![], Name::new("String"));
-        assert_eq!(user.builtin_alias(), None);
-        // A `baml` class without a primitive alias is not collapsed.
-        assert_eq!(baml_qtn(&[], "SomethingElse").builtin_alias(), None);
     }
 }
