+++ baml_language/crates/baml_type/Cargo.toml
@@ -7,3 +7,4 @@ rust-version.workspace = true
 [dependencies]
 baml_base = { workspace = true }
 borsh = { workspace = true }
+subenum = { workspace = true }
